+++ packages/components/src/base/Button/index.tsx
@@ -1,3 +1,1 @@
-export { Button } from './button';
-export { ButtonGroup } from './buttongroup';
-export { IconButton } from './iconbutton';
+export { Button } from './Button';
