+++ packages/components/src/base/ButtonGroup/index.tsx
@@ -1,0 +1,1 @@
+export { ButtonGroup } from './ButtonGroup';
