+++ packages/components/src/base/IconButton/index.tsx
@@ -1,0 +1,1 @@
+export { IconButton } from './IconButton';
