+++ packages/components/src/custom/Toolbar/custom-column.tsx
@@ -1,9 +1,9 @@
 import { ColumnIcon } from '@layer5/sistent-svg';
 import React from 'react';
-import { IconButton } from '../../base/Button';
 import { Checkbox } from '../../base/Checkbox';
 import { ClickAwayListener } from '../../base/ClickAwayListener';
 import { FormControlLabel } from '../../base/Form';
+import { IconButton } from '../../base/IconButton';
 import { Paper } from '../../base/Paper';
 import { Popper } from '../../base/Popper';
 import { Tooltip } from '../../base/Tooltip';
