+++ packages/components/src/index.tsx
@@ -3,6 +3,7 @@ export * from './base/AppBar';
 export * from './base/Avatar';
 export * from './base/Box';
 export * from './base/Button';
+export * from './base/ButtonGroup';
 export * from './base/Card';
 export * from './base/Checkbox';
 export * from './base/Chip';
@@ -13,6 +14,7 @@ export * from './base/Divider';
 export * from './base/Drawer';
 export * from './base/Form';
 export * from './base/Grid';
+export * from './base/IconButton';
 export * from './base/Input';
 export * from './base/InputLabel';
 export * from './base/List';
